Added event for Scandinavian Karelia
When Muscovy or Russia control Karelia Scandinavia gets an event fired that gives them the choice between a Casus Belli on Karelia with immediate declaration of war for 5.5 Infamy or Lose 1 Infamy and lose 5 Prestige. When choosing to go to war Scandinavia when controlling Karelia also later get the state cored via a decision.