added decision to get Karelia cored
After having conquered Karelia through the event that fires when either Muscovy or Russia control Karelia Scandinavia gets the decision to get the state of Karelia cored. Conditions for that are that Scandinavia is a GP, has Nationalism and Imperialism Researched, controls Karelia and has the country_flag given by the already mentioned event.