Several events added for flavor to ARI
First event is the Occitania Conquered event which fires if all Occitanian cores are owned by Aragon-Italy. Will set the occitania_conquered flag which will enable the Integrate Occitania decision to appear in the decisions menu. 

Second event is the Macedonian question, similar to the Status of Greece. Fires when Macedonia exists and then offers to transfer the Greek provinces to Macedonia in exchange for vassalizing Macedonia. Done to add flavor to the fact Macedonia being initiated and lead by a cadet branch of the Aragonese Royal Family.

Third event is the transfer of the Status of Greece Event from Italy to Aragon-Italy. Done as it made sense to have the event from Italy also for Aragon-Italy.